refactor: split interactive mode into modular components
